requirements.txt
connectonion
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
pydantic>=2.0.0
chromadb>=0.4.0
pypdf>=3.0.0
pytest>=7.0.0
httpx>=0.24.0
